requirements.txt
networkx==3.4.2
numpy==1.25.0
omegaconf==2.3.0
paretoset==1.2.4
pulp==3.1.1
pygmo==2.19.5
torch==2.0.1
tqdm==4.67.1
torch-geometric=2.6.1
matplotlib=3.10.1
hydra-core=1.3.2
